+++ Sources/Navigator/Toolkit/PaginationView.swift
@@ -29,9 +29,13 @@ enum PageLocation: Equatable {
     }
 }
 
+@MainActor
 protocol PageView {
     /// Moves the page to the given internal location.
     func go(to location: PageLocation, animated: Bool) async
+
+    /// Called when the page view is not the currently visible page anymore.
+    func pageDidDisappear()
 }
 
 @MainActor
@@ -209,6 +213,7 @@ final class PaginationView: UIView, Loggable {
         let movingBackward = (currentIndex - 1 == index)
         let location = location ?? (movingBackward ? .end : .start)
 
+        let previousView = currentView
         currentIndex = index
 
         // To make sure that the views the most likely to be visible are loaded first, we first load
@@ -225,6 +230,8 @@ final class PaginationView: UIView, Loggable {
                 continue
             }
         }
+
+        previousView?.pageDidDisappear()
 
         loadPages()
     }