+++ TestApp/Sources/Reader/Common/Preferences/UserPreferences.swift
@@ -100,7 +100,7 @@ struct UserPreferences<
                     )
 
                 case let editor as EPUBPreferencesEditor:
-                    switch editor.layout {
+                    switch editor.defaultLayout {
                     case .reflowable:
                         reflowableUserPreferences(
                             commit: commit,
@@ -786,7 +786,7 @@ struct UserPreferences<
             isActive: isActive,
             onClear: onClear
         ) {
-            HStack {
+            HStack(spacing: 4) {
                 Stepper(title,
                         onIncrement: onIncrement,
                         onDecrement: onDecrement)
@@ -861,7 +861,7 @@ struct UserPreferences<
         onClear: @escaping () -> Void,
         content: @escaping () -> V
     ) -> some View {
-        HStack {
+        HStack(spacing: 8) {
             content()
                 .foregroundColor(isActive ? nil : .gray)